move XWikiWebPreferences to correct package
