+++ celements-user-manager/component/src/main/java/com/celements/mandatory/XWikiWebPreferences.java
@@ -1,4 +1,4 @@
-package com.celements.migrator;
+package com.celements.mandatory;
 
 import java.util.List;
