+++ celements-user-manager/component/src/test/java/com/celements/mandatory/XWikiWebPreferencesTest.java
@@ -12,7 +12,6 @@ import org.xwiki.model.reference.DocumentReference;
 import org.xwiki.model.reference.WikiReference;
 
 import com.celements.common.test.AbstractComponentTest;
-import com.celements.migrator.XWikiWebPreferences;
 import com.celements.model.classes.ClassDefinition;
 import com.celements.model.classes.fields.ClassField;
 import com.celements.model.context.ModelContext;
